% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile-historical-rates.R
\name{compile_historical_rates}
\alias{compile_historical_rates}
\title{Aggregate historical interest rate data
For use with scrape_rates}
\usage{
compile_historical_rates()
}
\value{
A dataframe of monthly interest rates for all available years
}
\description{
Aggregate historical interest rate data
For use with scrape_rates
}
