% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download-and-clean.R
\name{download_and_clean}
\alias{download_and_clean}
\title{Download and clean interest rates
For use with compile_historical_rates}
\usage{
download_and_clean(url)
}
\arguments{
\item{url}{The URL from which to download the interest rate data}
}
\value{
A dataframe of monthly interest rates
}
\description{
Download and clean interest rates
For use with compile_historical_rates
}
