% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape-rates.R
\name{scrape_rates}
\alias{scrape_rates}
\title{Returns stabilized and nonstabilized segment rates for any year}
\usage{
scrape_rates()
}
\description{
Returns stabilized and nonstabilized segment rates for any year
}
