% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape-treasury-rates.R
\name{scrape_treasury_rates}
\alias{scrape_treasury_rates}
\title{Returns treasury rates for the given duration
Available columns to pull:
1 mo, 2 mo, 3 mo, 6 mo,
1 yr, 2 yr, 3 yr, 5 yr,
7 yr, 10 yr, 20 yr, 30 yr
Data is generally updated at the end of each business day}
\usage{
scrape_treasury_rates(year, column_of_interest)
}
\arguments{
\item{year}{The calendar year to pull results for}

\item{column_of_interest}{The duration of rates to pull info on}
}
\description{
Returns treasury rates for the given duration
Available columns to pull:
1 mo, 2 mo, 3 mo, 6 mo,
1 yr, 2 yr, 3 yr, 5 yr,
7 yr, 10 yr, 20 yr, 30 yr
Data is generally updated at the end of each business day
}
